+++ practica3/src/Gear.java
@@ -38,54 +38,54 @@ public abstract class Gear implements Soldier {
     /**
      * Method that returns the soldier´s damage points.
      * 
-     * @return int damage points
+     * @return double damage points
      */
-    public int getDp() {
+    public double getDp() {
         return soldier.getDp();
     }
 
     /**
      * Method that sets the soldier´s damage points.
      * 
-     * @param int dp representing damage points
+     * @param double dp representing damage points
      */
-    public void setDp(int dp){
+    public void setDp(double dp){
 	soldier.setDp(dp);
     }
 
     /**
      * Method that returns the soldier´s speed.
      * 
-     * @return int speed.
+     * @return double speed.
      */
-    public int getSpeed() {
+    public double getSpeed() {
         return soldier.getSpeed();
     }
 
     /**
      * Method that sets the soldier´s speed points.
      * 
-     * @param int speed representing speed points
+     * @param double speed representing speed points
      */
-    public void setSpeed(int speed){
+    public void setSpeed(double speed){
 	soldier.setSpeed(speed);
     }
 
     /**
      * Method that returns the soldier´s defense points.
      * 
-     * @return int defense points
+     * @return double defense points
      */
-    public int getDefense() {
+    public double getDefense() {
         return soldier.getDefense();
     }
 
     /**
      * Method that sets the soldier´s defense points.
      * 
-     * @param int defense representing defense points
+     * @param double defense representing defense points
      */
-    public void setDefense(int defense){
+    public void setDefense(double defense){
 	soldier.setDefense(defense);
     }
