+++ practica3/src/Soldier.java
@@ -22,44 +22,44 @@ public interface Soldier {
     /**
      * Method that returns the soldier´s damage points.
      * 
-     * @return int damage points
+     * @return double damage points
      */
-    public int getDp();
+    public double getDp();
 
     /**
      * Method that sets the soldier´s damage points.
      * 
-     * @param int dp representing damage points
+     * @param double dp representing damage points
      */
-    public void setDp(int dp);
+    public void setDp(double dp);
 
     /**
      * Method that returns the soldier´s speed.
      * 
-     * @return int speed.
+     * @return double speed.
      */
-    public int getSpeed();
+    public double getSpeed();
 
     /**
      * Method that sets the soldier´s speed points.
      * 
-     * @param int speed representing speed points
+     * @param double speed representing speed points
      */
-    public void setSpeed(int speed);
+    public void setSpeed(double speed);
 
     /**
      * Method that returns the soldier´s defense points.
      * 
-     * @return int defense points
+     * @return double defense points
      */
-    public int getDefense();
+    public double getDefense();
 
     /**
      * Method that sets the soldier´s defense points.
      * 
-     * @param int defense representing defense points
+     * @param double defense representing defense points
      */
-    public void setDefense(int defense);
+    public void setDefense(double defense);
 
     /**
      * Method that sets the soldier´s gear.
